ipc finally working

+++ ipc/ipc.cpp
@@ -16,14 +16,12 @@ bool IPC::_send_data(int fifo, std::unique_ptr<char[]> buffer, uint64_t buffer_s
     if(write(fifo, &buffer_size, sizeof(uint64_t)) != sizeof(uint64_t) ||
         write(fifo, buffer.get(), buffer_size) != buffer_size) 
         return false;
-
     return true;
 }
 
 std::unique_ptr<char[]> IPC::_receive_data(int fifo, uint64_t &buffer_size) {
     buffer_size = 0;
-    if(read(fifo, &buffer_size, sizeof(uint64_t)) != sizeof(uint64_t))
-        return nullptr;
+    while(read(fifo, &buffer_size, sizeof(uint64_t)) != sizeof(uint64_t));
 
     auto buffer = std::make_unique<char[]>(buffer_size);
 
@@ -58,8 +56,6 @@ std::unique_ptr<FIFO> FIFO::create_ipc(const char *rd_fifo_path, const char *wr_
     return std::move(std::make_unique<FIFO>(rd_fifo_pth, wr_fifo_pth, rd_fifo, wr_fifo));
 
 cleanup:
-    unlink(rd_fifo_path);
-    unlink(wr_fifo_path);
     free(rd_fifo_pth);
     free(wr_fifo_pth);
 
@@ -77,12 +73,10 @@ std::unique_ptr<char[]> FIFO::receive_data(uint64_t &buffer_size) {
 FIFO::~FIFO() {
     close(m_rd_fifo);
     close(m_wr_fifo);
-
-    unlink(m_rd_path.get());
-    unlink(m_wr_path.get());
 }
 
-SOCKT::SOCKT(char *socket_path, int socket) : m_socket(socket) {
+SOCKT::SOCKT(char *socket_path, bool is_server, int server_socket, int client_socket) 
+        : m_server(is_server), m_server_socket(server_socket), m_client_socket(client_socket) {
     m_socket_path = std::unique_ptr<char>(socket_path);
 }
 
@@ -90,19 +84,22 @@ std::unique_ptr<SOCKT> SOCKT::create_ipc(const char *_socket_path, const char *h
     char *socket_path = new char[strlen(_socket_path)];
     sockaddr_un socket_addr, client_addr;
     int socket_fd = -1, len = -1, client_socket = -1;
+    bool is_server = !strcmp(host, "server");
 
     strcpy(socket_path, _socket_path);
         
     // create a socket
-    if((socket_fd = socket(AF_UNIX, SOCK_STREAM, 0)) < 0)
+    if((socket_fd = socket(AF_UNIX, SOCK_STREAM, 0)) < 0) {
+        perror("SOCKET CREATE: ");
         goto onerror;
+    }
         
     // set the socket address structure to point to socket file
     memset(&socket_addr, 0, sizeof(socket_addr));
     socket_addr.sun_family = AF_UNIX;
     strncpy(socket_addr.sun_path, socket_path, sizeof(socket_addr.sun_path) - 1);
 
-    if(!strcmp(host, "server")) {
+    if(is_server) {
         // bind socket to socket file
         len = sizeof(socket_addr.sun_family) + strlen(socket_addr.sun_path);
         if(bind(socket_fd, (sockaddr *)(&socket_addr), len) < 0) {
@@ -111,39 +108,47 @@ std::unique_ptr<SOCKT> SOCKT::create_ipc(const char *_socket_path, const char *h
         }
         
         // listen on the socket
-        if(listen(socket_fd, 1000) < 0)
+        if(listen(socket_fd, 100000) < 0) {
+            perror("LISTEN : ");
             goto onerror;
+        }
 
-        if((client_socket = accept(socket_fd, nullptr, nullptr)) < 0)
+        return std::move(std::make_unique<SOCKT>(socket_path, is_server, socket_fd, -1));
+    } else if(!is_server) {
+        while(connect(socket_fd, (sockaddr *) &socket_addr, sizeof(socket_addr)) < 0) {
+            perror("CONNECT: ");
             goto onerror;
-
-        close(socket_fd);
-        return std::move(std::make_unique<SOCKT>(socket_path, client_socket));
-    } else if(!strcmp(host, "client")) {
-        if(connect(socket_fd, (sockaddr *) &socket_addr, sizeof(socket_addr)) < 0)
-            goto onerror;
+        }
         
-        return std::move(std::make_unique<SOCKT>(socket_path, socket_fd));
+        return std::move(std::make_unique<SOCKT>(socket_path, is_server, -1, socket_fd));
     }
 
 onerror:
-    unlink(socket_path);
+    if(is_server) 
+        unlink(socket_path);
     free(socket_path);
-    close(socket_fd);
     return nullptr;   
 }
 
 SOCKT::~SOCKT() {
-    close(m_socket);
-    unlink(m_socket_path.get());
+    if(m_server) 
+        unlink(m_socket_path.get());
 }
 
 bool SOCKT::send_data(std::unique_ptr<char[]> buffer, uint64_t buffer_size) {
-    return _send_data(m_socket, std::move(buffer), buffer_size);
+
+    bool output = _send_data(m_client_socket, std::move(buffer), buffer_size);
+    return output;
 }
 
 std::unique_ptr<char[]> SOCKT::receive_data(uint64_t &buffer_size) {
-    return _receive_data(m_socket, buffer_size);
+    if(!m_server)
+        return _receive_data(m_client_socket, buffer_size);
+    
+    close(m_client_socket);
+    if((m_client_socket = accept(m_server_socket, nullptr, nullptr)) < 0)
+        perror("ACCEPT : ");
+    return _receive_data(m_client_socket, buffer_size);
 }
 
 std::unique_ptr<SHM> SHM::create_ipc(const char *shm_path, const char *host) {
@@ -187,7 +192,8 @@ SHM::~SHM() {
     if(m_shmbuf)
         munmap(m_shmbuf, sizeof(*m_shmbuf));
 
-    shm_unlink(m_shm_path.get());
+    if(m_server)
+        shm_unlink(m_shm_path.get());
     m_shmbuf = nullptr;
 }
